Experiment in giving click callbacks access to the game
This works (although I don't quite understand the gymnastics around the move closure), and atm just attempts a flatstone placement move on the local game data, followed by printing the state out.

This points at the main problem this runs into: After changing the board state, we need to update the display. For this we currently need a ref to Client itself, but we can't hand out a pointer to client from within client. Also we have no place to store Client inside an RcRefCell.
Additionally, updating the display drops the Closure from which we'd call the display function in the first place. Do Closure objects store the rust closure on the heap, like an Rc? Otherwise this seems very dubious.