Experiment to give the click callback access to update the display
This is awful code.
I made the display type the full Rc<RefCell so that it can clone itself and pass the cloned Rc into the click closure.
However I feel that I need to re-order the whole situation. I don't think I can avoid using an Rc<RefCell entirely, but this wrapping feels silly. It also requires weird gymnastics like scoping the mutable game ref in on_click so that it gets dropped before we actually call update_display, since that borrows mutably again.
Also, update_display only needs to mutate the closure list, so really I'd want to be able to get finer-grained mutable access to just the game / just the display data (aka the closure list).
Might be that I can implement all the functionality directly on what is now DisplayData, but pass an instance of the Rc<RefCell of that same data through the update_display function and the capture closure?