Restructure client/display divide and communication
The new architecture makes the display responsible for updating visuals and registering input events,
and moves all functionality that responds to inputs and events into the Client.
The client is effectively the controller + data store (the game), and the display generates the view of that data.

Any input callback needs a (usually mutable) ref to the client to send events, which requires the client to be stored inside an Rc RefCell. The display also needs an instance of this Rc to hand out to the callbacks it generates.
The display is only accessed by the client currently, so the client owns it directly.

This means that the two structs circularly reference each other. Setup for this is handled by the init function now, since we need to construct each struct in sequence, and then patch up a reference (in this case, hand the display to the client struct).

The display storing an instance of the Rc RefCell also means that after the client is put behind that reference, it does not need to know about how it is stored anymore. This means that we can avoid implementing on a wrapper type of the Rc RefCell, just so that we can hand an instance of it to other code via self.clone(). This also means that the implementation of Client functionality can stay directly on the Client object, meaning that its methods' &self/&mut self parameters correctly signal what they do. (As opposed to methods on an Rc RefCell type, which may as well always take &self but might mutate the underlying data).

The Js side still calls a few methods directly on the Client. Previously, it "owned" the instance of the Client itself so those were directly bindgen'ed. However, now the Client instance lives behind the Rc RefCell, so we need to hand Js an instance of that instead. To do so and still expose the same interface, we have the ClientInterface wrappere type over the Rc RefCell, which just forwards the relevant method calls to the internal client object.
This has the added benefit of making it super obvious which methods are actually exposed to Js.