all; Send history and add history viewing
Server now only sends the game history + the board size, which fully defines the game state.
Server keeps a MoveHistory locally to apply moves to.
Client constructs a MoveHistory each time it gets a new move list from the server.
On click on any of the past moves, the client shows the state after that move.
While in that state, the client can't make any changes to the board.
If it's their turn, they can click on the space corresponding to their move in progress to get back to the interactable state.
This is probably bad UX (need to click on an invisible space to get back to being able to play). The upside is that it doesn't obliterate
your in-progress move if you go back in history, but this might be a weird edge case I shouldn't care about.
Will need future consideration.