server&client; First pass at random game ID generation
Increments a game id counter on the server on root connect, then turns that into a game ID and redirects to that sub-url.
The sub-url then gives out the index.html.
On the client side, the game id is taken from the url and passed to the /ws path.
That now parses the hash and will eventually pass that to the connection handler to connect the player to the correct match.

Note that the redirect is kind of wasteful and incorrect here (have to use the temp redirect, otherwise it gets cached).
Since the GET response doesn't care about what the game id is, only the WS connection does, we should really just send the game ID
to the client in the response, then have the client re-write the browser URL and pass the hash along when making the WS connection.