client; Connection in Rust, Fix Rc cycle, LateInit
Moved the connection into Rust, directly using the web-sys WebSocket type.
Got rid of the corresponding JS code, which means now all logic lives in Rust!

As a result of there not being any JS-side objects referencing the Client, it now needs a static reference in Rust code to prevent it from getting dropped as soon as our start function exits.
This didn't actually happen initially due to Rc reference cycles between the Client/Display and Client/Connection. Fixed this so that the Client owns Display/Connection, and they each only have a Weak reference to the Client object.

While I was changing that code, I also replaced the Client's `Option`s holding the Client/Connection with a new `LateInit` type, representing a value that is briefly uninitialised after being constructed, but is later initialised once to a value.
These now correctly express that a fully initialised Client struct will have valid values stored in those fields, and is more ergonomic due to LateInit auto-dereferencing to the inner value.